[JENKINS-19526] Branch Coverage Information
Instead of coverage based on the instruction counter, the displayed
partial and full coverage is based solely on branch information per
line of code. Lines are colored in green (full coverage), yellow
(partly covered) or red (not covered) for where coverage information
is available. Lines without coverage information are not formatted
any differently from the previous version.

Code with branch information is prefixed with a • and has a mouse-over
to display either "All X branches covered", "All X branches missed" or
"X of Y branches missed" for full, none and partial coverage
respectively.