Do not change job parameters with env variables
Environment variables when expanded changed the job settings.
It should only expand during the actual build, so that parameterized
build jobs don't have to be reverted after every run.

Moved expansion a bit earlier, so that the log messages contains
the patterns when they have been expanded.

Added a new testcase, that will fail without the changes to
JacocoPublisher.java.