Add option to remove original exec file
Build parameter that would trigger the removal of the
original exec file after it has been copied to the build
job directory.

The use-case being that an exec file that has been fetched
from a jacoco-agent can be quite huge, and would also only
be valid for when the job has finished. Being that a copy
exists in the build directory already.

Added test in JacocoPublisherTest::testGetterSetter, also
added a missing test for setBuildOverBuild/isBuildOverBuild.