Option to use symlinks for source and classes
The use case is a big application (more than
1 Gb, less than 2 gb), which collects metrics
via jacoco-agent every night, to get statitics
of functional test coverage.

When the source and classes is copied for each
job, the disk space usage get quite big.

When it's only the metrics changing, and not the
classes and source, it makes more sense to use
symbolic links for these in the job directory
instead of a copy.