Remove obsolete maven-enforcer ASM exclude
It was originally introduced because the ASM jar files contain a
module-info.class which is targeted to Java 9. When using JDK 8
for building the enforcer therefore complained.

The current parent pom uses the newer extra-enforcer-rules 1.2
which ignores the module-info.class in case a JDK < 9 is used.

For more details see
- https://github.com/mojohaus/extra-enforcer-rules/issues/33
- https://github.com/mojohaus/extra-enforcer-rules/issues/46