Suppress spotbugs PA_PUBLIC_PRIMITIVE_ATTRIBUTE warnings
https://www.jenkins.io/doc/developer/tutorial-improve/add-more-spotbugs-checks/
describes the technique used here with an exclusions file.  The specific
exclusion for PA_PUBLIC_PRIMITIVE_ATTRIBUTE has been confirmed by Jenkins
developers that it is not helpful in a Jenkins plugin.

https://github.com/jenkinsci/jenkins/pull/8803 describes the changes
that were made in Jenkins core and in many Jenkins plugins to suppress
the PA_PUBLIC_PRIMITIVE_ATTRIBUTE spotbugs warning.